pyproject.toml
[project]
name = "web-image-optimizer"
version = "0.1.0"
description = "CLI tool for reducing image size (JPEG/PNG/WebP) with resizing, compression, and parallel processing."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pillow",
    "rich",
]

[project.scripts]
wio = "wio:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
